Added Gameplay
Added obstacle spawner, bounded movement, a shooting mechanic, you can perish, and you can destroy objetcs. I added a few scripts for these things and I also changed the skybox. There is still a lot of refining to do and I need to review the code. This was done in a stupor at 2 in the morning.