Added Abilities, Different Menus, etc
I added a lot of stuff this time. I added three different abilities based on Zodiac Signs. I also added different menus for the title screen. I added a new packet which I mostly used for cool auras for the ship. A lot of code adding or reworking. Added a bit of story behind the game and also a gameplay explanation. I think that is it.